+++ fictionops/src/fictionops/core.py
@@ -5,6 +5,7 @@ from .agent_author_guards import *
 from .agent_counterevidence_review import *
 from .agent_counterevidence_apply import *
 from .agent_counterevidence_revision import *
+from .agent_counterevidence_candidate import *
 from .agent_evidence_escalation import *
 from .agent_failure_lab import *
 from .agent_connect import *